refactor: Improve component structure and add type definitions
- Refactored GitHubStats and ProjectShowcase components into smaller, more manageable sub-components.
- Created dedicated directories for these components and their sub-components (e.g., src/components/GitHubStats/, src/components/ProjectShowcase/).
- Added TypeScript interfaces for the data structures used within these components to improve type safety and code clarity.
- Updated all relevant import paths.
