+++ src/components/GitHubStats/ContributionChart.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { LineChart, Line, XAxis, YAxis, ResponsiveContainer, Tooltip } from 'recharts';
+
+interface ContributionItem { // Ensure this interface matches or is imported
+  month: string;
+  contributions: number;
+}
+
+interface ContributionChartProps {
+  data: ContributionItem[];
+}
+
+const ContributionChart: React.FC<ContributionChartProps> = ({ data }) => {
+  return (
+    <Card>
+      <CardHeader>
+        <CardTitle>Monthly Contributions</CardTitle>
+      </CardHeader>
+      <CardContent>
+        <ResponsiveContainer width="100%" height={250}>
+          <LineChart data={data}>
+            <XAxis dataKey="month" />
+            <YAxis />
+            <Tooltip />
+            <Line
+              type="monotone"
+              dataKey="contributions"
+              stroke="#22c55e"
+              strokeWidth={3}
+              dot={{ fill: '#22c55e', strokeWidth: 2, r: 4 }}
+            />
+          </LineChart>
+        </ResponsiveContainer>
+      </CardContent>
+    </Card>
+  );
+};
+
+export default ContributionChart;
