+++ src/components/ProjectShowcase/ProjectCard.tsx
@@ -1,0 +1,61 @@
+import React from 'react';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/ui/button';
+import { Github } from 'lucide-react'; // Assuming ExternalLink was not used directly for live demo
+
+interface Project { // Ensure this interface matches or is imported from index.tsx
+  title: string;
+  description: string;
+  tech: string[];
+  github: string;
+  live?: string;
+  status: string;
+  featured: boolean;
+}
+
+interface ProjectCardProps {
+  project: Project;
+}
+
+const ProjectCard: React.FC<ProjectCardProps> = ({ project }) => {
+  return (
+    <Card className={`card-hover ${project.featured ? 'ring-2 ring-primary/20' : ''}`}>
+      <CardHeader>
+        <div className="flex items-start justify-between">
+          <CardTitle className="text-lg">{project.title}</CardTitle>
+          <Badge variant={project.status === 'Live' ? 'default' : 'secondary'}>
+            {project.status}
+          </Badge>
+        </div>
+        <p className="text-sm text-muted-foreground">{project.description}</p>
+      </CardHeader>
+      <CardContent className="space-y-4">
+        <div className="flex flex-wrap gap-2">
+          {project.tech.map((tech, techIndex) => (
+            <Badge key={techIndex} variant="outline" className="text-xs">
+              {tech}
+            </Badge>
+          ))}
+        </div>
+        <div className="flex space-x-2">
+          <Button size="sm" variant="outline" asChild>
+            <a href={project.github} target="_blank" rel="noopener noreferrer">
+              <Github size={16} className="mr-1" />
+              Code
+            </a>
+          </Button>
+          {project.live && (
+            <Button size="sm" asChild>
+              <a href={project.live} target="_blank" rel="noopener noreferrer">
+                Live Demo
+              </a>
+            </Button>
+          )}
+        </div>
+      </CardContent>
+    </Card>
+  );
+};
+
+export default ProjectCard;
