+++ src/pages/Index.tsx
@@ -4,8 +4,8 @@ import Header from '@/components/Header';
 import StatsOverview from '@/components/StatsOverview';
 import DifficultyChart from '@/components/DifficultyChart';
 import ContestStats from '@/components/ContestStats';
-import GitHubStats from '@/components/GitHubStats';
-import ProjectShowcase from '@/components/ProjectShowcase';
+import GitHubStats from '@/components/GitHubStats/index';
+import ProjectShowcase from '@/components/ProjectShowcase/index';
 import LiveProjects from '@/components/LiveProjects';
 import Now from '@/components/Now';
 import Footer from '@/components/Footer';
